pyproject.toml
[project]
name = "recoco-sync"
version = "0.2.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "django-celery-results>=2.5.1",
    "django-environ>=0.11.2",
    "django-extensions>=3.2.3",
    "django-model-utils>=5.0.0",
    "django-ninja>=1.3.0",
    "django>=5.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.27.2",
    "psycopg2-binary>=2.9.9",
    "redis>=5.1.0",
    "sentry-sdk[celery,django]>=2.15.0",
    "whitenoise>=6.7.0",
    "django-browser-reload>=1.18.0",
    "django-debug-toolbar>=5.0.1",
]

[tool.uv]
dev-dependencies = [
    "factory-boy>=3.3.1",
    "ipython>=8.28.0",
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.6.8",
    "unittest-parametrize>=1.4.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "recoco_sync.settings.testing"
addopts = ["--reuse-db"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
show-fixes = true
target-version = "py312"
exclude = [".venv"]

[tool.ruff.lint]
select = [
    "F",   # Pyflakes
    "E",   # pycodestyle
    "W",   # pycodestyle
    "C90", # mccabe
    "I",   # isort
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "BLE", # flake8-blind-except
    "FBT", # flake8-boolean-trap
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "T20", # flake8-print
    "Q",   # flake8-quotes
    "TID", # flake8-tidy-imports
    "PGH", # pygrep-hooks
    "INP", # flake8-no-pep420
]
unfixable = ["B"]
ignore = [
    "ISC001", # Implicit string concatenation
]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401", "F403"]
"sandbox/*" = ["T201", "T203"]
"**/settings/testing.py" = ["E402", "F403"]
"**/settings/dev.py" = ["E402", "F403", "F405"]
"**/migrations/*.py" = ["E501"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
case-sensitive = true
known-third-party = [
    "django",
    "environ",
    "factory",
    "model_utils",
    "ninja",
    "unittest_parametrize",
]
split-on-trailing-comma = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
required-imports = ["from __future__ import annotations"]

[tool.ruff.format]
skip-magic-trailing-comma = false
